Integrate callback into the eventManager, configure sandbox callback
