+++ sandbox/public/index.html
@@ -51,7 +51,13 @@
         logEnabled: true,
         prehidingStyle: ".personalization-container { opacity: 0 !important }",
         idSyncContainerId: 81,
-        migrateIds: true
+        migrateIds: true,
+        onBeforeEventSend(xdm) {
+          const urlSearchParams = new URLSearchParams(window.location.search);
+          if (urlSearchParams.has("title")) {
+            xdm.web.webPageDetails.name = urlSearchParams.get("title");
+          }
+        }
       });
 
       alloy("event", {
