+++ src/components/DataCollector/index.js
@@ -10,8 +10,6 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import createConfigValidators from "./createConfigValidators";
-
 const createDataCollector = ({ eventManager }) => {
   return {
     commands: {
@@ -41,6 +39,6 @@ const createDataCollector = ({ eventManager }) => {
 
 createDataCollector.namespace = "DataCollector";
 createDataCollector.abbreviation = "DC";
-createDataCollector.configValidators = createConfigValidators();
+createDataCollector.configValidators = {};
 
 export default createDataCollector;
