+++ src/core/createConfigValidators.js
@@ -10,23 +10,29 @@ OF ANY KIND, either express or implied. See the License for the specific languag
 governing permissions and limitations under the License.
 */
 
-import { string, callback } from "../../utils/configValidators";
+import { boolean, string, callback } from "../utils/configValidators";
 
-export default () => {
-  return {
-    configId: {
-      validate: string().unique()
-    },
-    edgeDomain: {
-      defaultValue: "beta.adobedc.net",
-      validate: string().domain()
-    },
-    imsOrgId: {
-      validate: string().unique()
-    },
-    onBeforeEventSend: {
-      defaultValue: () => undefined,
-      validate: callback()
-    }
-  };
-};
+export default () => ({
+  errorsEnabled: {
+    validate: boolean(),
+    defaultValue: true
+  },
+  logEnabled: {
+    validate: boolean(),
+    defaultValue: false
+  },
+  configId: {
+    validate: string().unique()
+  },
+  edgeDomain: {
+    defaultValue: "beta.adobedc.net",
+    validate: string().domain()
+  },
+  imsOrgId: {
+    validate: string().unique()
+  },
+  onBeforeEventSend: {
+    defaultValue: () => undefined,
+    validate: callback()
+  }
+});
